Move testing helpers into their own folder
