+++ tests/Actions/MarkAsShippedTest.php
@@ -5,7 +5,7 @@ namespace DoubleThreeDigital\SimpleCommerce\Tests\Actions;
 use DoubleThreeDigital\SimpleCommerce\Actions\MarkAsShipped;
 use DoubleThreeDigital\SimpleCommerce\Facades\Order;
 use DoubleThreeDigital\SimpleCommerce\Facades\Product;
-use DoubleThreeDigital\SimpleCommerce\Tests\SetupCollections;
+use DoubleThreeDigital\SimpleCommerce\Tests\Helpers\SetupCollections;
 use DoubleThreeDigital\SimpleCommerce\Tests\TestCase;
 use Statamic\Facades\Collection;
 use Statamic\Facades\Entry;
