+++ tests/Helpers/UseDatabaseContentDrivers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DoubleThreeDigital\SimpleCommerce\Tests;
+namespace DoubleThreeDigital\SimpleCommerce\Tests\Helpers;
 
 use Illuminate\Support\Facades\File;
 
@@ -8,7 +8,7 @@ trait UseDatabaseContentDrivers
 {
     public function setUpDatabaseContentDrivers()
     {
-        $this->stubsPath = __DIR__ . '/__fixtures__/database/migrations';
+        $this->stubsPath = __DIR__ . '/../__fixtures__/database/migrations';
 
         if (count(File::glob(database_path('migrations') . '/*_create_customers_table.php')) < 1) {
             File::copy($this->stubsPath . '/create_customers_table.php', database_path('migrations/' . date('Y_m_d_His') . '_create_customers_table.php'));
